Implemented context data
Context data is similar to defining an environment in LaTeX, so that macros in Jinja can act as pseudo-environments. This allows one to define a piece of data for a context, and macros in subsequent calls()'s for nested macro calls can utilize this. This is useful for things such as maintaining a counter or keeping data for the duration of a macro call.