Implemented external module loading.
External module loading has been implemented. Paths are prefixed with an @ symbol, and are more easily worked with. This system automatically worked with  jinja's structure, and can do things, such as

{% include "@MyModule.SubModule/template.jinja" %}

This will search for the template subdirectory in the python module MyModule.SubModule, and create a file directory loader from there.